agregar/quitar usuarios de capacitacnion (aplicado a backend)
